# contracts/v1.3/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_restore_backup_api_backup_restore_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "target_dir": {
            "title": "Target Dir",
            "type": "string"
          }
        },
        "required": [
          "file",
          "target_dir"
        ],
        "title": "Body_restore_backup_api_backup_restore_post",
        "type": "object"
      },
...
      "ExportReq": {
        "properties": {
          "scope": {
            "default": "project",
            "title": "Scope",
            "type": "string"
          },
          "unit_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Id"
          }
        },
        "title": "ExportReq",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LocalBackupRestoreReq": {
        "properties": {
          "backup_path": {
            "title": "Backup Path",
            "type": "string"
          },
          "target_dir": {
            "title": "Target Dir",
            "type": "string"
          }
        },
        "required": [
          "backup_path",
          "target_dir"
        ],
        "title": "LocalBackupRestoreReq",
        "type": "object"
      },
...
      "RecoveryPointRestoreReq": {
        "properties": {
          "recovery_point_id": {
            "title": "Recovery Point Id",
            "type": "string"
          },
          "target_dir": {
            "title": "Target Dir",
            "type": "string"
          }
        },
        "required": [
          "recovery_point_id",
          "target_dir"
        ],
        "title": "RecoveryPointRestoreReq",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "审迹",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/backup/create": {
      "post": {
        "description": "备份项目（audit.db + 附件库）到上级目录 .auditbak。",
        "operationId": "create_backup_api_backup_create_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Backup"
      }
    },
    "/api/backup/download/{filename}": {
      "get": {
        "description": "下载备份 .auditbak（存放于项目上级目录，不走输出目录端点）。\n\n防目录穿越：解析后的路径必须落在项目上级目录内。",
        "operationId": "download_backup_api_backup_download__filename__get",
        "parameters": [
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Backup"
      }
    },
    "/api/backup/recovery-point": {
      "post": {
        "description": "手工立即创建一份增量恢复点；仍使用已保存的自动备份目标和空间策略。",
        "operationId": "create_auto_recovery_point_api_backup_recovery_point_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Auto Recovery Point"
      }
    },
    "/api/backup/recovery-points": {
      "get": {
        "description": "列出当前项目自动备份目标中的可用恢复点。",
        "operationId": "list_auto_recovery_points_api_backup_recovery_points_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Auto Recovery Points"
      }
    },
    "/api/backup/recovery-points/restore": {
      "post": {
        "description": "从内容寻址自动备份恢复点恢复；始终写入一个新项目目录。",
        "operationId": "restore_auto_recovery_point_api_backup_recovery_points_restore_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryPointRestoreReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Auto Recovery Point"
      }
    },
    "/api/backup/restore": {
      "post": {
        "description": "恢复备份：上传 .auditbak + 目标目录（须为空）。",
        "operationId": "restore_backup_api_backup_restore_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_restore_backup_api_backup_restore_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Backup"
      }
    },
    "/api/backup/restore-local": {
      "post": {
        "description": "从本机路径恢复完整 .auditbak，避免 50GB 文件经浏览器上传的大小限制。",
        "operationId": "restore_local_backup_api_backup_restore_local_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalBackupRestoreReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Local Backup"
      }
    },
    "/api/backup/settings": {
      "get": {
        "operationId": "get_backup_settings_api_backup_settings_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Backup Settings"
      },
      "post": {
        "operationId": "save_backup_settings_api_backup_settings_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupSettingsReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Backup Settings"
      }
    },
    "/api/diagnostics/support-package": {
      "post": {
        "description": "生成默认剔除业务内容和路径的本机诊断支持包。",
        "operationId": "export_diagnostics_support_package_api_diagnostics_support_package_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Diagnostics Support Package"
      }
    },
...
    "/api/export/excel": {
      "post": {
        "description": "导出问题汇总表 Excel（unit=当前单位 / project=全部单位）。",
        "operationId": "export_excel_api_export_excel_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Excel"
      }
    },
    "/api/export/file/{filename}": {
      "get": {
        "description": "下载输出目录中的导出文件（Excel/ZIP/备份）。",
        "operationId": "download_export_api_export_file__filename__get",
        "parameters": [
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Export"
      }
    },
    "/api/export/package": {
      "post": {
        "description": "通过归档核对令牌后打包 ZIP；项目变化或核对过期必须重新确认。",
        "operationId": "package_project_api_export_package_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackageReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Package Project"
      }
    },
    "/api/export/package/preflight": {
      "post": {
        "description": "生成归档核对清单。无阻断项时发放一次性确认令牌，有效期 10 分钟。",
        "operationId": "package_preflight_api_export_package_preflight_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackageReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Package Preflight"
      }
    },
...
    "/api/import/excel": {
      "post": {
        "description": "上传整理好的 xlsx，一键导入底稿（单位不存在自动创建）。",
        "operationId": "import_excel_api_import_excel_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_excel_api_import_excel_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Excel"
      }
    },
    "/api/import/excel/commit": {
      "post": {
        "description": "提交与预检摘要一致的 Excel；底稿写入为整批原子替换。",
        "operationId": "commit_excel_import_api_import_excel_commit_post",
        "parameters": [
          {
            "in": "query",
            "name": "confirmation_token",
            "required": false,
            "schema": {
              "default": "",
              "title": "Confirmation Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_commit_excel_import_api_import_excel_commit_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commit Excel Import"
      }
    },
    "/api/import/excel/preflight": {
      "post": {
        "description": "只读校验 Excel，显示影响摘要后发放一次性提交令牌。",
        "operationId": "import_excel_preflight_api_import_excel_preflight_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_excel_preflight_api_import_excel_preflight_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Excel Preflight"
      }
    },
    "/api/import/merge": {
      "post": {
        "description": "旧上传入口已停用：正式合并需本机路径预检，避免大包限制和绕过冲突确认。",
        "operationId": "import_merge_api_import_merge_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_merge_api_import_merge_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Merge"
      }
    },
    "/api/import/merge-local": {
      "post": {
        "description": "通过预检确认后从本机路径合并，适用于 50GB 场景。",
        "operationId": "import_merge_local_api_import_merge_local_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalMergeReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Merge Local"
      }
    },
    "/api/import/merge-local/preflight": {
      "post": {
        "description": "对本机备份来源做只读预检，发现冲突先展示，确认后才允许写入。",
        "operationId": "import_merge_local_preflight_api_import_merge_local_preflight_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-session",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Session",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalMergeReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Merge Local Preflight"
      }
    },
...
  }
}